Pacify MSVC (firxed some warnings related to type conversions)

+++ Arrangement_on_surface_2/demo/Arrangement_on_surface_2_earth/Camera.cpp
@@ -17,7 +17,7 @@ Camera::Camera() :
 {}
 
 //! \brief
-void Camera::perspective(float fov, float aspect, float z_near, float z_far) {
+void Camera::perspective(qreal fov, qreal aspect, qreal z_near, qreal z_far) {
   m_z_near = z_near;
   m_z_far = z_far;
